add source map sample

+++ scripts/generate-website.js
@@ -2,7 +2,7 @@ const fs = require('fs-extra')
 const path = require('path')
 
 const root = path.join(__dirname, '..')
-const samples = ['minimal-sample', 'web-worker-sample']
+const samples = ['minimal-sample', 'web-worker-sample', 'source-map-sample']
 
 fs.ensureDirSync(path.join(root, 'website'))
 fs.ensureDirSync(path.join(root, 'website/samples'))
@@ -17,8 +17,10 @@ for (const sample of samples) {
     path.join(root, `packages/html-preview-web-samples/${sample}/dist`),
     path.join(root, `website/samples/${sample}/dist`)
   )
-  fs.removeSync(path.join(root, `website/samples/${sample}/dist/remoteMain.js`))
-  fs.removeSync(path.join(root, `website/samples/${sample}/dist/remoteMain.js.map`))
+  if (!['source-map-sample'].includes(sample)) {
+    fs.removeSync(path.join(root, `website/samples/${sample}/dist/remoteMain.js`))
+    fs.removeSync(path.join(root, `website/samples/${sample}/dist/remoteMain.js.map`))
+  }
 }
 
 const html = `
